Fix compile error after change to use C++23
list of compile error

/home/hpichu/work/dali-adaptor/dali/internal/adaptor/common/combined-update-render-controller.cpp: In member function ‘virtual void Dali::Internal::Adaptor::CombinedUpdateRenderController::ResizeSurface()’:
/home/hpichu/work/dali-adaptor/dali/internal/adaptor/common/combined-update-render-controller.cpp:423:5: error: ‘++’ expression of ‘volatile’-qualified type is deprecated [-Werror=volatile]
  423 |     mSurfaceResized++;

/home/hpichu/work/dali-adaptor/dali/internal/accessibility/bridge/bridge-object.cpp: In lambda function:
/home/hpichu/work/dali-adaptor/dali/internal/accessibility/bridge/bridge-object.cpp:90:167: error: implicit capture of ‘this’ via ‘[=]’ is deprecated in C++20 [-Werror=deprecated]
   90 |     AddCoalescableMessage(static_cast<CoalescableMessages>(static_cast<int>(CoalescableMessages::PROPERTY_CHANGED_BEGIN) + static_cast<int>(event)), obj.get(), 1.0f, [=, weakObj = std::weak_ptr<Accessible>(obj)]()
      |                                                                                                                                                                       ^
/home/hpichu/work/dali-adaptor/dali/internal/accessibility/bridge/bridge-object.cpp:90:167: note: add explicit ‘this’ or ‘*this’ capture

/home/hpichu/work/dali-adaptor/automated-tests/src/dali-adaptor/utc-Dali-FileLoader.cpp:69:28: error: ambiguous overload for ‘operator==’ (operand types are ‘std::streampos’ {aka ‘std::fpos<__mbstate_t>’} and ‘int’)
   69 |   DALI_TEST_CHECK(fileSize == 0);

+++ plugins/dali-feedback.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Samsung Electronics Co., Ltd.
+ * Copyright (c) 2026 Samsung Electronics Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -69,7 +69,7 @@ void DaliFeedback::StopHaptic()
 int DaliFeedback::PlaySound(const std::string& fileName)
 {
   int handle    = -1;
-  int errorCode = mm_sound_play_keysound(fileName.c_str(), VOLUME_TYPE_SYSTEM & VOLUME_GAIN_TOUCH);
+  int errorCode = mm_sound_play_keysound(fileName.c_str(), static_cast<int>(VOLUME_TYPE_SYSTEM) & static_cast<int>(VOLUME_GAIN_TOUCH));
   if(errorCode < 0)
   {
     DEBUG_PRINTF("PlaySound() %s failed with error code = %d\n", fileName.c_str(), errorCode);
